+++ controllers/index.js
@@ -2,9 +2,13 @@ const router = require('express').Router()
 
 const apiRoutes = require('./api')
 const homeRoutes = require('./home-routes.js')
+const loginRoutes = require('./login')
+const signupRoutes = require('./signup')
 // const dashboardRoutes = require('./dashboard-routes.js')
 
 router.use('/api', apiRoutes)
+router.use('/login', loginRoutes)
+router.use('/signup', signupRoutes)
 router.use('/', homeRoutes)
 // router.use ('/dashboard', dashboardRoutes)
