+++ controllers/login/index.js
@@ -1,0 +1,3 @@
+const routes = require('./login-routes');
+
+module.exports = routes;
