+++ controllers/signup/index.js
@@ -1,0 +1,3 @@
+const routes = require('./signup-routes');
+
+module.exports = routes;
